pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-dev-team-platform"
version = "0.1.0"
description = "Reusable AI development team orchestration platform for Slack, Discord, and future web dashboards."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "pydantic>=2.7.0,<3.0.0",
    "pydantic-settings>=2.2.1,<3.0.0",
    "pyyaml>=6.0.1,<7.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0,<1.0.0",
    "pytest>=8.2.0,<9.0.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["apps*", "packages*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
